Bring PICK_WINNER routine definition up to date with production version
Setting the winner_id datatype as smallint was creating issue where if
the winner_id was greater than 32767 then the user with that id --
32767, i.e. Nadine Bale -- would be selected the winner. Apparently this
was fixed at some point in production but not updated in the repo.

+++ doc/routines.sql
@@ -163,7 +163,7 @@ PICK_WINNER_SPROC:BEGIN
     DECLARE existing_contest TINYINT(1) DEFAULT NULL;
     DECLARE existing_winner_id,
             new_winner_user_id,
-            new_winner_site_id SMALLINT(5) DEFAULT NULL;
+            new_winner_site_id BIGINT(10) DEFAULT NULL;
 
     -- check if the contest date exists and if a winner has already been chosen
     SELECT
